LibWasm: Store opcodes in 32 bits
Wasm opcodes only need one byte plus a 24-bit selector for
prefixed instructions. Store them in u32 instead of u64 and reject
selectors that would not fit. Update the Rust opcode generator to
accept the new integer suffixes used by Opcode.h.

+++ Libraries/LibWasm/CraneliftBridge.cpp
@@ -746,7 +746,7 @@ static CraneliftInsn serialize_insn(Dispatch const& dispatch, SourcesAndDestinat
     out.imm3 = 0;
 
     auto const& args = insn->arguments();
-    u64 opc = out.opcode;
+    u32 opc = out.opcode;
 
     if (opc == Instructions::i32_const.value()) {
         out.imm1 = static_cast<i64>(args.get<i32>());
